+++ README.md
@@ -1,6 +1,6 @@
 # PSR-7 & PSR-15 router library
 
-`codeinc/router` is a [PSR-15](https://www.php-fig.org/psr/psr-15/) router library written in PHP 7, processing [PSR-7](https://www.php-fig.org/psr/psr-7/) requests and responses. A router is a [PSR-15](https://www.php-fig.org/psr/psr-15/) middleware implementing [`MiddlewareInterface`](https://www.php-fig.org/psr/psr-15/#22-psrhttpservermiddlewareinterface) in charge of determining which request handler to call to answer a request, to call the selected controller and then to returns the PSR-7 response. All routers must implement the [`RouterInterface`](src/RouterInterface.php) interface. 
+`codeinc/router` is a [PSR-15](https://www.php-fig.org/psr/psr-15/) router library written in PHP 7, processing [PSR-7](https://www.php-fig.org/psr/psr-7/) requests and responses. A router is a [PSR-15](https://www.php-fig.org/psr/psr-15/) middleware implementing [`MiddlewareInterface`](https://www.php-fig.org/psr/psr-15/#22-psrhttpservermiddlewareinterface) in charge of determining which request handler to call to answer a request, to call the selected controller and then to returns the PSR-7 response. All routers must implement the [`RouterInterface`](src/ResolverInterface.php) interface. 
 
 Two types of routers are supplied in the current package: a static router which takes a list of predefined routes and request handlers classes and a dynamic router which computes the routes of the request handlers base on the request handler's namespace (compatible with [PSR-0](https://www.php-fig.org/psr/psr-0/) or [PSR-4](https://www.php-fig.org/psr/psr-4/) naming conventions). Both are available as `final` ([`StaticRouter`](src/StaticRouter.php) and [`DynamicRouter`](src/DynamicRouter/DynamicRouter.php)) and `abstract` classes ([`AbstractStaticRouter`](src/StaticRouter/AbstractStaticRouter.php) and [`AbstractDynamicRouter`](src/DynamicRouter/DynamicRouter.php)).
