--- src/DynamicRouter/DynamicRouter.php
@@ -1,82 +1,0 @@
-<?php
-//
-// +---------------------------------------------------------------------+
-// | CODE INC. SOURCE CODE                                               |
-// +---------------------------------------------------------------------+
-// | Copyright (c) 2018 - Code Inc. SAS - All Rights Reserved.           |
-// | Visit https://www.codeinc.fr for more information about licensing.  |
-// +---------------------------------------------------------------------+
-// | NOTICE:  All information contained herein is, and remains the       |
-// | property of Code Inc. SAS. The intellectual and technical concepts  |
-// | contained herein are proprietary to Code Inc. SAS are protected by  |
-// | trade secret or copyright law. Dissemination of this information or |
-// | reproduction of this material is strictly forbidden unless prior    |
-// | written permission is obtained from Code Inc. SAS.                  |
-// +---------------------------------------------------------------------+
-//
-// Author:   Joan Fabrégat <joan@codeinc.fr>
-// Date:     24/09/2018
-// Project:  Router
-//
-declare(strict_types=1);
-namespace CodeInc\Router\DynamicRouter;
-use CodeInc\Router\RouterException;
-
-
-/**
- * Class DynamicRouter
- *
- * @package CodeInc\Router\DynamicRouter
- * @author Joan Fabrégat <joan@codeinc.fr>
- */
-class DynamicRouter extends AbstractDynamicRouter
-{
-    /**
-     * @var string
-     */
-    private $controllersNamespace;
-
-    /**
-     * @var string
-     */
-    private $uriPrefix;
-
-    /**
-     * DynamicRouter constructor.
-     *
-     * @param string $controllersNamespace
-     * @param string $uriPrefix
-     * @throws RouterException
-     */
-    public function __construct(string $controllersNamespace, string $uriPrefix)
-    {
-        if (empty($uriPrefix)) {
-            throw RouterException::emptyUriPrefix();
-        }
-        if (empty($controllersNamespace)) {
-            throw RouterException::emptyControllersNamespace();
-        }
-        $this->controllersNamespace = $controllersNamespace;
-        $this->uriPrefix = $uriPrefix;
-    }
-
-    /**
-     * Returns the router's URI prefix.
-     *
-     * @return string
-     */
-    public function getUriPrefix():string
-    {
-        return $this->uriPrefix;
-    }
-
-    /**
-     * Returns the requests handler's base namespace.
-     *
-     * @return string
-     */
-    public function getControllersNamespace():string
-    {
-        return $this->controllersNamespace;
-    }
-}
